Switch to gopher-lua as more complete.
